chore: FINALLY fix v9 build
Previously, ngnes failed because it was unable to find
the node_modules folder.

I noticed that NES_NPM_CONFIG_LOCAL_PREFIX was being set to './'. This
is not done on any other branch.

It appears that this was left during a dubugging session. I deleted
that config setting and ngnes ran successfully.